requirements.txt
ansible==11.3.0
ansible-core==2.18.3
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
binaryornot==0.4.4
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
cryptography==44.0.2
decorator==5.2.1
Django==5.1.7
django-appconf==1.1.0
django-browser-reload==1.18.0
django-compressor==4.5.1
django-filter==24.3
django-json-widget==2.0.1
django-jsonform==2.23.2
django-tailwind==3.8.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
executing==2.2.0
gunicorn==23.0.0
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
openpyxl==3.1.5
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
rcssmin==1.1.2
requests==2.32.3
resolvelib==1.0.1
rich==13.9.4
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
urllib3==2.3.0
wcwidth==0.2.13
